Перенёс список и карту на главной странице в отдельный компонент
