+++ project/src/pages/main/main.tsx
@@ -1,22 +1,13 @@
-import {useState} from 'react';
 import {Link} from 'react-router-dom';
 import {Offer} from '../../types/offer';
-import OffersList from '../../components/offers-list/offers-list';
 import Logo from '../../components/logo/logo';
-import Map from '../../components/map/map';
+import PlacesContainer from '../../components/places-container/places-container';
 
 type MainComponentProps = {
   offers: Offer[];
 };
 
-const tabIndexValue = 0;
-
 function Main({offers}: MainComponentProps): JSX.Element {
-  const [selectedOffer, setSelectedOffer] = useState<Offer | undefined>(undefined);
-  const onOfferItemHover = (offerId: number): void => {
-    const currentOffer = offers.find((offer) => offer.id === offerId);
-    setSelectedOffer(currentOffer);
-  };
   return (
     <>
       <div style={{display: 'none'}}>
@@ -90,31 +81,7 @@ function Main({offers}: MainComponentProps): JSX.Element {
             </section>
           </div>
           <div className="cities">
-            <div className="cities__places-container container">
-              <section className="cities__places places">
-                <h2 className="visually-hidden">Places</h2>
-                <b className="places__found">{offers.length} places to stay in Amsterdam</b>
-                <form className="places__sorting" action="#" method="get">
-                  <span className="places__sorting-caption">Sort by</span>
-                  <span className="places__sorting-type" tabIndex={tabIndexValue}>
-                  Popular
-                    <svg className="places__sorting-arrow" width="7" height="4">
-                      <use xlinkHref="#icon-arrow-select"></use>
-                    </svg>
-                  </span>
-                  <ul className="places__options places__options--custom places__options--opened">
-                    <li className="places__option places__option--active" tabIndex={tabIndexValue}>Popular</li>
-                    <li className="places__option" tabIndex={tabIndexValue}>Price: low to high</li>
-                    <li className="places__option" tabIndex={tabIndexValue}>Price: high to low</li>
-                    <li className="places__option" tabIndex={tabIndexValue}>Top rated first</li>
-                  </ul>
-                </form>
-                <OffersList listType={'main'} offers={offers} onOfferItemHover={onOfferItemHover}/>
-              </section>
-              <div className="cities__right-section">
-                <Map className={'cities__map map'} offers={offers} selectedOffer={selectedOffer}/>
-              </div>
-            </div>
+            <PlacesContainer offers={offers}/>
           </div>
         </main>
       </div>
